Calvin 20240331 memento DP redo

+++ irentstuffapp/tests/test_views.py
@@ -9,7 +9,9 @@ from irentstuffapp.models import Item, Category, Message, Rental
 from irentstuffapp.forms import ItemForm, ItemEditForm, RentalForm
 from PIL import Image
 from unittest.mock import patch
+from django.core.exceptions import ValidationError
 import io
+import os
 
 
 class InboxViewTestCase(TestCase):
@@ -345,6 +347,15 @@ class AddItemViewTestCase(TestCase):
         self.assertEqual(Item.objects.first().owner, self.user)  # Check that item owner is correct
         self.assertRedirects(response, reverse("item_detail", kwargs={"item_id": Item.objects.first().id}))
 
+    def tearDown(self) -> None:
+        # Get the path to the image file
+        image_path = Item.objects.first().image.path
+
+        # Check if the file exists and delete it
+        if os.path.exists(image_path):
+            os.remove(image_path)
+        return super().tearDown()
+
 
 class EditItemViewTestCase(TestCase):
     def create_image(self, name="test_image.jpg", size=(1, 1), image_mode="RGB", image_format="JPEG"):
@@ -374,6 +385,7 @@ class EditItemViewTestCase(TestCase):
             created_date=datetime(2024, 2, 7, tzinfo=timezone.utc),
             deleted_date=None,
         )
+        self.image1_path = self.item.image.path
 
     def test_edit_item_authenticated_owner(self):
         # Login as the item owner
@@ -414,6 +426,16 @@ class EditItemViewTestCase(TestCase):
         self.assertEqual(self.item.deposit, 100.00)
         self.assertIn(self.image2.name.split(".")[0], self.item.image.name)
 
+    def tearDown(self):
+        # Get the paths to the image files
+        image2_path = self.item.image.path
+
+        # Check if the files exist and delete them
+        if os.path.exists(image2_path):
+            os.remove(image2_path)        
+        if os.path.exists(self.image1_path):
+            os.remove(self.image1_path)
+
 
 class DeleteItemViewTestCase(TestCase):
     def setUp(self):
@@ -440,13 +462,37 @@ class DeleteItemViewTestCase(TestCase):
         initial_item_count = Item.objects.count()
 
         # Make a POST request to delete_item view
-        response = self.client.post(reverse("delete_item", kwargs={"item_id": self.item.id}))
+        response = self.client.post(reverse("delete_item", kwargs={"item_id": self.item.id}), {'delete_confirm': 'confirmed'})
 
         # Check that the item was successfully deleted
         self.assertEqual(response.status_code, 302)  # Redirect after successful deletion
         self.assertEqual(Item.objects.count(), initial_item_count - 1)  # Check if item count is decreased
         self.assertIsNone(Item.objects.filter(pk=self.item.id).first())  # Check if item is deleted from database
 
+    def test_delete_item_with_rental(self): 
+        self.renter = User.objects.create_user(username="testrenter", email="rent@example.com", password="password321")
+
+        self.rental = Rental.objects.create(
+            renter=self.renter,
+            owner=self.user,
+            item=self.item,
+            start_date=date.today(),
+            end_date=date.today() + timedelta(days=3),
+            status="confirmed",
+        )
+        # Login as the user
+        self.client.login(username="testuser", password="password123")
+
+        # Get the initial count of items
+        initial_item_count = Item.objects.count()
+
+        # Make a POST request to delete_item view
+        response = self.client.post(reverse("delete_item", kwargs={"item_id": self.item.id}))
+
+        # Check that the item was successfully deleted
+        self.assertEqual(response.status_code, 302)
+        self.assertEqual(Item.objects.count(), initial_item_count)  # Check if item count did not decrease
+        self.assertIsNotNone(Item.objects.filter(pk=self.item.id).first()) 
 
 class ItemDetailViewTestCase(TestCase):
     def setUp(self):
@@ -835,3 +881,5 @@ class ItemMessagesViewTestCase(TestCase):
 
         # Check if the response status code is 200 (OK)
         self.assertEqual(response.status_code, 200)
+
+
